refactor asyncraft path
