+++ Cargo.toml
@@ -73,7 +73,7 @@ bytes = "1"
 byteorder = "1.4"
 tracing = "0.1.29"
 tracing-subscriber = { version = "0.3.0", features = ["env-filter"] }
-reqwest = { version = "0.11", features = ["json"], default-features = false }
+# reqwest = { version = "0.11", features = ["json"], default-features = false }
 async-raft = "0.6"
 thiserror = "1.0.20"
 clap = { version = "4.3", features = ["derive"] }
