+++ src/raft/cluster/mod.rs
@@ -6,7 +6,7 @@ use crate::{common::appdata::AppShareData, config::core::{ConfigKey, ConfigAsync
 
 use self::model::{RouterRequest, RouterResponse};
 
-use super::{asyncraft::store::ClientRequest, join_node};
+use super::{store::ClientRequest, join_node};
 
 pub mod route;
 pub mod routeapi;
