+++ src/raft/mod.rs
@@ -1,13 +1,15 @@
 use async_raft::{Raft, RaftStorage};
 
-use self::asyncraft::{store::{ClientRequest, ClientResponse, store::AStore}};
-use self::asyncraft::network::network::RaftRouter;
+use self::{store::{ClientRequest, ClientResponse, store::RaftStore}};
+use self::network::network::RaftRouter;
 
-pub mod asyncraft;
 pub mod cluster;
-pub type NacosRaft = Raft<ClientRequest, ClientResponse, RaftRouter, AStore>;
+pub mod store;
+pub mod network;
 
-pub async fn join_node(raft:&NacosRaft,raft_store:&AStore,node_id:u64) -> anyhow::Result<()> {
+pub type NacosRaft = Raft<ClientRequest, ClientResponse, RaftRouter, RaftStore>;
+
+pub async fn join_node(raft:&NacosRaft, raft_store:&RaftStore, node_id:u64) -> anyhow::Result<()> {
     let membership=raft_store.get_membership_config().await?;
     if !membership.contains(&node_id) {
         let mut all_node=  membership.all_nodes();
